feat: CardList 가 데이터를 받아서 화면을 그리도록 수정
